details: show common data even on unknown item kinds
This means that a new item kind will sort-of work, even if it isn't
finished.
