+++ src/test/java/somethingrandom/usecase/details/ItemDetailsFailureTest.java
@@ -2,10 +2,8 @@ package somethingrandom.usecase.details;
 
 import org.junit.Before;
 import org.junit.Test;
-import somethingrandom.entity.Item;
 import somethingrandom.usecase.DataAccessException;
 
-import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -14,17 +12,6 @@ import static org.junit.Assert.assertTrue;
 
 public class ItemDetailsFailureTest {
     private DummyPresenter presenter;
-
-    private static class FakeItem extends Item {
-        public FakeItem(String name, UUID id, Instant creationDate) {
-            super(name, id, creationDate);
-        }
-
-        @Override
-        public String getItemKind() {
-            return "FAKE";
-        }
-    }
 
     @Before
     public void setup() {
@@ -47,14 +34,5 @@ public class ItemDetailsFailureTest {
 
         usecase.showDetailsForId(UUID.fromString("634f6f85-7cb8-4af5-9cc3-17fec4dcf0b6"));
         assertTrue(presenter.message.contains("Exception Message"));
-    }
-
-    @Test
-    public void shouldPresentFailureWithUnknownItemKind() {
-        Item item = new FakeItem("Fake", UUID.fromString("634f6f85-7cb8-4af5-9cc3-17fec4dcf0b6"), Instant.MIN);
-        ItemDetailsInputBoundary usecase = new ItemDetailsInteractor(uuid -> Optional.of(item), presenter);
-
-        usecase.showDetailsForId(UUID.fromString("634f6f85-7cb8-4af5-9cc3-17fec4dcf0b6"));
-        assertNotNull(presenter.message);
     }
 }
